various stuff (PLEASE RESET ANY TEST DATABASES)

+++ src/entities/networks.py
@@ -12,7 +12,6 @@ class Network:
     def __init__(
         self,
         _id: str,
-        ip_address: str,
         country: str = None,
         asn: str = None,
         vpn: bool = False,
@@ -21,8 +20,7 @@ class Network:
         first_used: datetime = None,
         last_used: datetime = None
     ):
-        self.id = _id
-        self.ip_address = ip_address
+        self.ip_address = _id
         self.country = country
         self.asn = asn
         self.vpn = vpn
@@ -34,7 +32,6 @@ class Network:
     @property
     def admin(self):
         return {
-            "id": self.id,
             "ip_address": self.ip_address,
             "country": self.country,
             "asn": self.asn,
@@ -48,67 +45,63 @@ class Network:
 
     @property
     def users(self):
-        return [users.get_user(netlog["user_id"]) for netlog in db.netlog.find({"ip_address": self.ip_address})]
+        return [users.get_user(netlog["_id"]["user_id"]) for netlog in db.netlog.find({"_id.ip_address": self.ip_address})]
 
     @property
     def user_ids(self):
-        return [netlog["user_id"] for netlog in db.netlog.find({"ip_address": self.ip_address})]
+        return [netlog["_id"]["user_id"] for netlog in db.netlog.find({"_id.ip_address": self.ip_address})]
 
     def set_block_state(self, blocked: bool):
         self.blocked = blocked
-        db.networks.update_one({"_id": self.id}, {"$set": {"blocked": self.blocked}})
+        db.networks.update_one({"_id": self.ip_address}, {"$set": {"blocked": self.blocked}})
     
     def set_creation_block_state(self, creation_blocked: bool):
         self.creation_blocked = creation_blocked
-        db.networks.update_one({"_id": self.id}, {"$set": {"creation_blocked": self.creation_blocked}})
+        db.networks.update_one({"_id": self.ip_address}, {"$set": {"creation_blocked": self.creation_blocked}})
 
     def update_last_used(self):
         self.last_used = uid.timestamp()
-        db.networks.update_one({"_id": self.id}, {"$set": {"last_used": self.last_used}})
+        db.networks.update_one({"_id": self.ip_address}, {"$set": {"last_used": self.last_used}})
     
     def delete(self):
-        db.networks.delete_one({"_id": self.id})
+        db.networks.delete_one({"_id": self.ip_address})
 
 class Netlog:
     def __init__(
         self,
         _id: str,
-        ip_address: str = None,
-        user_id: str = None,
         first_used: datetime = None,
         last_used: datetime = None
     ):
-        self.id = _id
-        self.network = get_network(ip_address)
-        self.user = users.get_user(user_id)
+        self.user = users.get_user(_id["user_id"])
+        self.network = get_network(_id["ip_address"])
         self.first_used = first_used
         self.last_used = last_used
 
     @property
     def admin(self):
         return {
-            "id": self.id,
-            "network": self.network.admin,
             "user": self.user.partial,
+            "network": self.network.admin,
             "first_used": int(self.first_used.timestamp()),
             "last_used": int(self.last_used.timestamp())
         }
-
-    def update_last_used(self):
-        self.last_used = uid.timestamp()
-        db.netlog.update_one({"_id": self.id}, {"$set": {"last_used": self.last_used}})
     
+    def update(self):
+        self.network.update_last_used()
+        self.last_used = uid.timestamp()
+        db.netlog.update_one({"_id": {"user_id": self.user.id, "ip_address": self.network.ip_address}}, {"$set": {"last_used": self.last_used}})
+
     def delete(self):
-        db.netlog.delete_one({"_id": self.id})
+        db.netlog.delete_one({"_id": {"user_id": self.user.id, "ip_address": self.network.ip_address}})
 
 def get_network(ip_address: str):
-    network = db.networks.find_one({"ip_address": ip_address})
+    network = db.networks.find_one({"_id": ip_address})
     if network is None:
         if IPHUB_KEY:
             iphub_info = requests.get(f"https://v2.api.iphub.info/ip/{ip_address}", headers={"X-Key": IPHUB_KEY}).json()
             network = {
-                "_id": uid.snowflake(),
-                "ip_address": ip_address,
+                "_id": ip_address,
                 "country": iphub_info["countryName"],
                 "asn": iphub_info["asn"],
                 "vpn": (iphub_info["block"] == 1),
@@ -117,8 +110,7 @@ def get_network(ip_address: str):
             }
         else:
             network = {
-                "_id": uid.snowflake(),
-                "ip_address": ip_address,
+                "_id": ip_address,
                 "country": None,
                 "asn": None,
                 "vpn": False,
@@ -128,9 +120,9 @@ def get_network(ip_address: str):
         db.networks.insert_one(network)
     return Network(**network)
 
-def get_netlog(user: any, network: Network):
+def get_netlog(user_id: str, ip_address: str):
     # Get netlog from database
-    netlog = db.netlog.find_one({"user_id": user.id, "ip_address": network.ip_address})
+    netlog = db.netlog.find_one({"_id": {"user_id": user_id, "ip_address": ip_address}})
     
     # Return netlog object
     if netlog:
@@ -138,24 +130,21 @@ def get_netlog(user: any, network: Network):
     else:
         raise status.resourceNotFound
 
-def get_all_netlogs(user: any):
-    return [Netlog(**netlog) for netlog in db.netlog.find({"user_id": user.id})]
+def get_all_netlogs(user_id: str):
+    return [Netlog(**netlog) for netlog in db.netlog.find({"_id.user_id": user_id})]
 
-def update_netlog(user: any, network: Network):
+def update_netlog(user_id: str, ip_address: str):
     try:
-        netlog = get_netlog(user, network)
+        netlog = get_netlog(user_id, ip_address)
     except status.resourceNotFound:
         netlog = {
-            "_id": uid.snowflake(),
-            "user_id": user.id,
-            "ip_address": network.ip_address,
+            "_id": {"user_id": user_id, "ip_address": ip_address},
             "first_used": uid.timestamp(),
             "last_used": uid.timestamp()
         }
         db.netlog.insert_one(netlog)
+        netlog = Netlog(**netlog)
     else:
-        db.netlog.update_one({"_id": netlog["_id"]}, {"$set": {
-            "ip_address": network.ip_address,
-            "last_used": uid.timestamp()
-        }})
-    network.update_last_used()
+        netlog.update()
+    
+    return netlog
